==> web_page_analysis/1_preprocess.ipynb <==
### 1. Предобработка заголовков, извлечение текста страниц из html-кода

import nltk
import numpy as np
import pandas as pd
from nltk.corpus import stopwords
from pymystem3 import Mystem
from string import punctuation
mystem = Mystem() 
russian_stopwords = stopwords.words("russian")
stop_tokens = []

def is_word(token):
    legal_chars = '0123456789abcdefghijklmnopqrstuvwxyzабвгдеёжзийклмнопрстуфхцчшщъыьэюя'
    for c in legal_chars:
        if c in token:
            return True
    stop_tokens.append(token)
    return False

def preprocess_text(text):
    tokens = mystem.lemmatize(text.lower())
    tokens = [token.strip() for token in tokens if token not in russian_stopwords\
              and token != " " \
              and token.strip() not in punctuation
              and is_word(token)]
    
    text = " ".join(tokens)
    
    return text

doc_to_title = {}
with open('docs_titles.tsv') as f:
    for num_line, line in enumerate(f):
        if num_line == 0:
            continue
        data = line.strip().split('\t', 1)
        doc_id = int(data[0])
        if len(data) == 1:
            title = ''
        else:
            title = data[1]
        doc_to_title[doc_id] = preprocess_text(title)

# stop_tokens

import os
from html.parser import HTMLParser

contents = ''

class Parser(HTMLParser):
    def handle_data(self, data):
        global contents
        tag = self.get_starttag_text()
        if tag is not None and tag.find('script') == tag.find('style') == -1:
            contents += data.strip() + ' '

directory = !pwd
directory = directory[0]
os.makedirs(directory + '/content/preprocessed')

parser = Parser()

for dirname, _, filenames in os.walk(directory + '/content'):
    if dirname.find('preprocessed') != -1: #начнет залезать в созданную папку preprocessed, надо пресечь
        break
    for filename in filenames:
        file = os.path.join(dirname,filename)
        with open(file, 'r') as f:
            strings = f.readlines()
            strings = [s.strip() for s in strings[1:] ] #убираем адрес сайта, находящийся на 1 строке
            text = '\n'.join(strings)
            parser.feed(text)
            
            newfile = file[:file.rfind('/')] + '/preprocessed/' + filename[:-4] + '_preprocessed.dat'
            
            with open(newfile, 'w') as f1:
                print(contents, file = f1)
            print(len(contents), 'chars in', newfile)
            parser.reset()
            contents = ''

df = pd.read_csv('train_groups.csv')
df['title'] = pd.Series(dtype='object')
for i, row in df.iterrows():
    df['title'][i] = doc_to_title[row['doc_id']]
df.to_csv('train_preprocessed.csv', index=False)
df

df = pd.read_csv('test_groups.csv')
df['title'] = pd.Series(dtype='object')
for i, row in df.iterrows():
    df['title'][i] = doc_to_title[row['doc_id']]
df.to_csv('test_preprocessed.csv', index=False)
df
