==> web_page_analysis/2_fulltext_normalizer.ipynb <==
### 2. Предобработка текстов веб-страниц

import os
import nltk
import numpy as np
import pandas as pd
from nltk.corpus import stopwords
from pymystem3 import Mystem
from string import punctuation
mystem = Mystem() 
russian_stopwords = stopwords.words("russian")

def preprocess_text(text):
    def is_word(token):
        legal_chars = '0123456789abcdefghijklmnopqrstuvwxyzабвгдеёжзийклмнопрстуфхцчшщъыьэюя'
        for c in legal_chars:
            if c in token:
                return True
        return False

    tokens = mystem.lemmatize(text.lower())
    tokens = [token.strip() for token in tokens if token not in russian_stopwords
              and len(token) > 1
              and token.strip() not in punctuation
              and is_word(token)]
    
    text = " ".join(tokens)
    
    return text

def preprocess_file(file_name):
    with open(os.path.join('/Users/tiptop/Downloads/preprocessed', file_name)) as in_:
        out_name = file_name[:file_name.find('_')]
        with open(f'fulltext/{out_name}.txt', 'w') as out_:
            out_.write(preprocess_text(in_.read()))

from tqdm.notebook import tqdm
from multiprocessing import Pool

files = sorted(os.listdir('/Users/tiptop/Downloads/preprocessed'))
with Pool(4) as p:
    pages = list(tqdm(p.imap(preprocess_file, files), total=len(files)))
