==> web_page_analysis/4_model.ipynb <==
### 4. 
### TF-IDF преобразование нормализованных заголовков
### Построение модели с TF-IDF фичами заголовков и полных текстов
### Обучение и предсказание

import numpy as np
import pandas as pd
import warnings
import xgboost as xgb
from sklearn.ensemble import GradientBoostingClassifier, RandomForestClassifier
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from sklearn.metrics import f1_score
from sklearn.metrics.pairwise import cosine_distances
from sklearn.model_selection import GroupKFold
from sklearn.neighbors import NearestNeighbors
from sklearn.pipeline import make_pipeline
from sklearn.preprocessing import StandardScaler
from sklearn.svm import SVC
from tqdm.notebook import tqdm
from xgboost import XGBClassifier

def get_top_words(corpus, top_word_count = 5):
    vec = CountVectorizer().fit(corpus)
    bag_of_words = vec.transform(corpus)
    sum_words = bag_of_words.sum(axis=0)
    words_freq = [(word, sum_words[0, idx]) for word, idx in vec.vocabulary_.items()]
    words_freq = sorted(words_freq, key = lambda x: x[1], reverse=True)
    top_words = [w[0] for w in words_freq[:top_word_count]]
    return top_words
    

def add_features(df, tf_idf_feature_count=15):
    def scale(col_name):
        scaler = StandardScaler()
        scaler.fit(df[col_name].values.reshape(-1, 1))
        df[col_name] = scaler.transform(df[col_name].values.reshape(-1, 1))

    df['len_word'] = list(map(lambda x: len(x.split()), df['title']))
    # средняя длина заголовка в категории
    df['mean_len'] = np.zeros(len(df.index))
    df['var_len'] = np.zeros(len(df.index))
    for i in df['group_id'].unique():
        df.loc[df['group_id'] == i, 'mean_len'] = df.loc[df['group_id'] == i, 'len_word'].mean()
        df.loc[df['group_id'] == i, 'mean_len'] = df.loc[df['group_id'] == i, 'len_word'].var()
    scale('len_word')
    scale('mean_len')
    scale('var_len')
    # расстояния до ближайших соседей по tf-idf
    for j in range(tf_idf_feature_count):
        df[f'tif_{j}'] = np.zeros(len(df.index))
    knn = NearestNeighbors(metric='cosine')
    df['dist_mean'] = np.zeros(len(df.index))
    df['dist_var'] = np.zeros(len(df.index))
    for i in df['group_id'].unique():
        X = TfidfVectorizer().fit_transform(df.loc[df['group_id'] == i, 'title'])
        knn.fit(X)
        distances = knn.kneighbors(n_neighbors=tf_idf_feature_count)[0]
        for j in range(tf_idf_feature_count):
            df.loc[df['group_id'] == i, f'tif_{j}'] = distances[:, j]
        # характеристики попарных расстояний
        cd = cosine_distances(X)
        df.loc[df['group_id'] == i, 'dist_mean'] = cd.mean()
        df.loc[df['group_id'] == i, 'dist_var'] = cd.var()
    scale('dist_mean')
    scale('dist_var')
    scale('text_len_word')
    scale('text_len_word_mean')
    scale('text_len_word_var')
    # характеристики по топ-словам
    for i in df['group_id'].unique():
        group = df.loc[df['group_id'] == i]
        top_words = get_top_words(group['title'])
        warnings.filterwarnings('ignore')
        group['tit_words'] = list(map(lambda x:x.split(), group['title']))
        warnings.filterwarnings('default')
        df.loc[df['group_id'] == i, 'top_in_tit'] = list(map(lambda x:sum([w in x for w in top_words]) / len(top_words), group['tit_words']))
        df.loc[df['group_id'] == i, 'tit_in_top'] = list(map(lambda x:sum([w in top_words for w in x]) / max(1, len(x)), group['tit_words']))
    # относительная доля различных символов в заголовке
    df2 = df.copy()
    df2['dig_cnt'] = list(map(lambda x:sum(c.isdigit() for c in x), df['title']))
    df2['cyr_cnt'] = list(map(lambda x:sum(c in 'абвгдеёжзийклмнопрстуфхцчшщъыьэюя' for c in x), df['title']))
    df2['lat_cnt'] = list(map(lambda x:sum(c in 'abcdefghijklmnopqrstuvwxyz' for c in x), df['title']))
    df2['tit_len'] = list(map(lambda x:len(x), df['title']))
    df['dig_prop'] = df2['dig_cnt'] / df2['tit_len']
    df['cyr_prop'] = df2['cyr_cnt'] / df2['tit_len']
    df['lat_prop'] = df2['lat_cnt'] / df2['tit_len']
    # относительные доли символов по группам
    for i in df['group_id'].unique():
        tot_title = ''.join(df.loc[df['group_id'] == i, 'title'])
        tit_len = len(tot_title)
        df.loc[df['group_id'] == i, 'dig_prop_gr'] = sum(c.isdigit() for c in tot_title) / tit_len
        df.loc[df['group_id'] == i, 'cyr_prop_gr'] = sum(c in 'абвгдеёжзийклмнопрстуфхцчшщъыьэюя' for c in tot_title) / tit_len
        df.loc[df['group_id'] == i, 'lat_prop_gr'] = sum(c in 'abcdefghijklmnopqrstuvwxyz' for c in tot_title) / tit_len

df = pd.read_csv('train_fulltext2.csv').fillna('')
#add_features(df)
#df = df.fillna(0)
df

def get_classifier(X_train, y_train, model_):
    if model_ == 'svm':
        clf = make_pipeline(StandardScaler(), SVC(gamma='auto', class_weight='balanced'))
    elif model_ == 'xgb':    
        clf = XGBClassifier()
    elif model_ == 'gb':
        clf = GradientBoostingClassifier(random_state=42)
    elif model_ == 'rf':
        clf = RandomForestClassifier(random_state=42)
    clf.fit(X_train, y_train)
    return clf

def get_model_score(X, y, model, n_splits=3):
    group_kfold = GroupKFold(n_splits=n_splits)
    scores = []
    for train_index, test_index in group_kfold.split(X, y, df['group_id']):
        X_train, X_test = X.iloc[train_index], X.iloc[test_index]
        y_train, y_test = y[train_index], y[test_index]
        clf = get_classifier(X_train, y_train, model)
        y_pred = clf.predict(X_test)
        scores.append(f1_score(y_test, y_pred))
    return np.mean(scores)

X = df.drop(['pair_id', 'doc_id', 'title', 'target', 'group_id'], axis=1)
y = df['target']
model_scores = []
for model in tqdm(['svm', 'gb', 'rf']):
    model_scores.append(get_model_score(X, y, model))
np.mean(model_scores)
# 0.7554901890005347

model_scores

model = 'svm'
df['train'] = 'train'
df_test = pd.read_csv('test_fulltext2.csv').fillna('')
df_test['train'] = 'test'
full_df = df.append(df_test)
add_features(full_df)
full_df = full_df.fillna(0)
df_test = full_df.loc[full_df['train'] == 'test'].drop(['train'], axis=1)
df = df.drop(['train'], axis=1)
X = df.drop(['pair_id', 'doc_id', 'title', 'target', 'group_id'], axis=1)
y = df['target']
clf = get_classifier(X, y, model)
df_test['target'] = clf.predict(df_test.drop(['pair_id', 'group_id', 'doc_id', 'title', 'target'], axis=1))
df_test[['pair_id', 'target']].to_csv(f'predict_full_text_{model}.csv', index=False)
